refactor: move confirmProduct method to ProductService
